Align featured image in centre
If the featured image is small, it is off centred in posts. This fixes that.

Just a side note, [HTML div align](https://www.w3schools.com/tags/att_div_align.asp) is outdated, so there should be a better way. However, `div align` still is supported so it works well.